+++ backend/canisters/community/impl/src/updates/c2c_join_community.rs
@@ -87,6 +87,9 @@ fn is_permitted_to_join(args: &Args, state: &RuntimeState) -> Result<Option<(Acc
                             ii_origin: vc.ii_origin.clone(),
                         }
                     }),
+                    referred_by_member: args
+                        .referred_by
+                        .map_or(false, |user_id| state.data.members.get_by_user_id(&user_id).is_some()),
                     now: state.env.now(),
                 },
             )