+++ backend/canisters/user/api/src/queries/mod.rs
@@ -16,4 +16,5 @@ pub mod public_profile;
 pub mod saved_crypto_accounts;
 pub mod search_messages;
 pub mod token_swap_status;
+pub mod token_swaps;
 pub mod updates;