+++ backend/canisters/user_index/impl/src/jobs/mod.rs
@@ -1,6 +1,7 @@
 use crate::RuntimeState;
 pub mod make_pending_payments;
 pub mod remove_from_online_users_canister;
+pub mod reset_leaderboard;
 pub mod submit_message_to_modclub;
 pub mod sync_events_to_local_user_index_canisters;
 pub mod sync_users_to_identity_canister;
@@ -15,4 +16,5 @@ pub(crate) fn start(state: &RuntimeState) {
     sync_users_to_storage_index::start_job_if_required(state);
     sync_users_to_identity_canister::start_job_if_required(state);
     upgrade_canisters::start_job_if_required(state);
+    reset_leaderboard::start_job_if_required(state);
 }